zoom blur

+++ src/utils.rs
@@ -1,7 +1,7 @@
-pub fn get_current_time() -> f64 {
+pub fn get_current_time() -> f32 {
     let window = web_sys::window().expect("should have a window in this context");
     let performance = window
         .performance()
         .expect("performance should be available");
-    performance.now()
+    performance.now() as f32
 }
